ADD #24 pokemon-ish font
