+++ web/index.css
@@ -1,9 +1,19 @@
+@font-face {
+    font-family: "PressStart2P";
+    src: url("assets/fonts/Press_Start_2P/PressStart2P-Regular.ttf");
+    font-weight: 400;
+}
+
 body {
     background-color: #333;
     color: #ffffffee;
 
     text-shadow: 0 0.05rem 0.1rem rgba(0, 0, 0, 0.5);
     box-shadow: inset 0 0 5rem rgba(0, 0, 0, 0.5);
+}
+
+body * {
+    font-family: "PressStart2P";
 }
 
 h1 {
