+++ web/index.html
@@ -33,7 +33,6 @@
             </div>
         </footer>
     </div>
-    <!-- TODO consider Pokemon style Font https://fonts.google.com/specimen/Press+Start+2P?preview.text=Glurak&preview.text_type=custom -->
 
     <script src="build/bundle.js"></script>
     <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.0.2/dist/js/bootstrap.bundle.min.js"
